# internal/transport/http/swaggerdocs/swagger.yaml
basePath: /api/v1
definitions:
...
  shared.Costume3DListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/shared.Costume3DObjectResponse'
        type: array
      pagination:
        $ref: '#/definitions/shared.PaginationResponse'
    type: object
  shared.Costume3DObjectResponse:
    properties:
      assetbundleName:
        type: string
      characterId:
        type: integer
      colorId:
        type: integer
      designer:
        type: string
      groupId:
        type: integer
      id:
        type: integer
      name:
        type: string
      partType:
        type: string
      publishedAt:
        type: integer
      rarity:
        type: string
      seq:
        type: integer
      type:
        type: string
    type: object
...
  shared.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/shared.ErrorBody'
    type: object
...
  shared.PaginationResponse:
    properties:
      has_next:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for master data sync and card querying.
  title: sekai-master-api
  version: "1.0"
paths:
...
  /costume3ds/{region}/{id}:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: 3D costume ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.Costume3DObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get a 3D costume by id
      tags:
      - costume3ds
  /costume3ds/{region}/list:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Case-insensitive substring of the normalized costume name
        in: query
        name: name
        type: string
      - description: Comma-separated group IDs
        in: query
        name: group_id
        type: string
      - description: Comma-separated color IDs
        in: query
        name: color_id
        type: string
      - description: Comma-separated character IDs
        in: query
        name: character_id
        type: string
      - description: Sort field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc|desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.Costume3DListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List 3D costumes by page
      tags:
      - costume3ds
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"